unlike

+++ eatraid/Backend/server.js
@@ -481,19 +481,22 @@ app.post("/add-to-fav", async (req, res) => {
 });
 
 app.delete("/delete-fav", async (req, res) => {
-  const { user, restaurant } = req.body;
+  const { restaurant } = req.body;
+  if (!req.session.userId) {
+    return res.status(401).json({ msg: "User not authenticated" });
+  } else {
+    const { error } = await supabase
+      .from('Favorite')
+      .delete()
+      .eq('UserId', req.session.userId)
+      .eq('RestaurantId', restaurant);
 
-  const { error } = await supabase
-    .from('Favorite')
-    .delete()
-    .eq('UserId', user)
-    .eq('RestaurantId', restaurant);
-
-  if (error) {
-    res.status(400).json(error);
-  }
-  else {
-    res.status(200).json({ 'msg': "delete user's fav restaurant successfully" });
+    if (error) {
+      res.status(400).json(error);
+    }
+    else {
+      res.status(200).json({ 'msg': "delete user's fav restaurant successfully" });
+    }
   }
 });
